Don't accept EULA by default and add mechanisms for the user to accept
Previously we were accepting the FSL EULA by default in the local.conf
file.

We changed this behavior by allowing the user to either accept
automatically by exporting an environment variable with
ACCEPT_EULA_${MACHINE_STR} pattern or by displaying the EULA in the
screen and capture the user confirmation manually.

The acceptance just needs to be done one time.

Fixes IOTMBL-1751: Fix EULAs in MBL build and for evaluation binaries